Actualizacion DB

+++ database/migrations/2021_08_17_114218_create_zonas_envio_table.php
@@ -14,7 +14,7 @@ class CreateZonasEnvioTable extends Migration
     public function up()
     {
         Schema::create('zonas_envios', function (Blueprint $table) {
-            $table->integer('id')->autoIncrement();
+            $table->bigInteger('id')->autoIncrement();
             $table->string('nombreZona', 60);
             $table->string('regionZona', 60);
             $table->string('metodosEnvio', 60);
